feat: add app setup
