+++ src/common/setup/setup.ts
@@ -1,0 +1,37 @@
+import { INestApplication, ValidationPipe } from '@nestjs/common';
+import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import { SWAGGER } from 'src/constants/enums/swagger/swagger';
+
+export function setupApp(app: INestApplication): void {
+  app.useGlobalPipes(
+    new ValidationPipe({
+      whitelist: true,
+      forbidNonWhitelisted: true,
+    }),
+  );
+
+  app.setGlobalPrefix('api');
+
+  app.enableCors({
+    origin: [
+      'http://localhost:5500',
+      'http://127.0.0.1:5500',
+      'http://localhost:3000',
+      'https://weather-api-application-tau.vercel.app',
+      'http://35.207.129.35:3000',
+    ],
+    credentials: true,
+    methods: 'GET,HEAD,PUT,PATCH,POST,DELETE,OPTIONS',
+    allowedHeaders: '*',
+  });
+
+  const config = new DocumentBuilder()
+    .setTitle(SWAGGER.TITLE)
+    .setDescription(SWAGGER.DESCRIPTION)
+    .setVersion(SWAGGER.VERSION)
+    .addTag(SWAGGER.TAG)
+    .build();
+
+  const document = SwaggerModule.createDocument(app, config);
+  SwaggerModule.setup('api/docs', app, document);
+}
