+++ src/main.ts
@@ -1,40 +1,13 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
 import { ensureLogDirExists } from 'src/utils/logger/logger.config';
-import { ValidationPipe } from '@nestjs/common';
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { SWAGGER } from './constants/enums/swagger/swagger';
+import { setupApp } from './common/setup/setup';
 
 ensureLogDirExists();
 
 async function bootstrap(): Promise<void> {
   const app = await NestFactory.create(AppModule);
-
-  app.useGlobalPipes(new ValidationPipe({ whitelist: true }));
-  app.setGlobalPrefix('api');
-
-  const config = new DocumentBuilder()
-    .setTitle(SWAGGER.TITLE)
-    .setDescription(SWAGGER.DESCRIPTION)
-    .setVersion(SWAGGER.VERSION)
-    .addTag(SWAGGER.TAG)
-    .build();
-
-  app.enableCors({
-    origin: [
-      'http://localhost:5500',
-      'http://127.0.0.1:5500',
-      'http://localhost:3000',
-      'https://weather-api-application-tau.vercel.app',
-      'http://35.207.129.35:3000',
-    ],
-    credentials: true,
-    methods: 'GET,HEAD,PUT,PATCH,POST,DELETE,OPTIONS',
-    allowedHeaders: '*',
-  });
-
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('api/docs', app, document);
+  setupApp(app);
 
   await app.listen(process.env.PORT ?? 3000);
 }
